requirements.txt
einops==0.7.0
pandas==1.5.3
peft==0.7.1
Pillow==10.2.0
scikit_learn==1.3.0
#torch==2.2.0.dev20231208+cu121
#torchvision==0.17.0.dev20231208+cu121
tqdm==4.65.0
transformers==4.38.0.dev0
